SMB items (messages or files) can now have 32-bit or 64-bit cost value
To fully support files > 4GB in size in file bases, credit values larger than 32-bits must be supported too.

There's a couple of todo comments/items included in this commit, but that's mainly to do with messages (which don't really have costs anyway).

The main thing to deal with now is the fact that users can't have more than 4GB in credits in the first place! That's got to be fixed next.

+++ src/sbbs3/atcodes.cpp
@@ -2056,7 +2056,7 @@ const char* sbbs_t::atcode(char* sp, char* str, size_t maxlen, long* pmode, bool
 			return str;
 		}
 		if(strcmp(sp, "FILE_CREDITS") == 0) {
-			safe_snprintf(str, maxlen, "%lu", (ulong)current_file->cost);
+			safe_snprintf(str, maxlen, "%" PRIu64, current_file->cost);
 			return str;
 		}
 		if(strcmp(sp, "FILE_TIME") == 0)
